+++ phpBB2/pafiledb/modules/pa_search.php
@@ -380,7 +380,7 @@ class pafiledb_search extends pafiledb_public
 
 					if ( $row = $db->sql_fetchrow($result) )
 					{
-						$search_data = phpbb_safe_unserialize($row['search_array']);
+						$search_data = phpbb_safe_unserialize_array($row['search_array']);
 						if (!is_array($search_data) || empty($search_data['pafiledb']))
 						{
 							message_die(GENERAL_MESSAGE, $lang['No_search_match']);